fix: rename contents domain and services to _domain and _services to avoid redundant content script generation in Plasmo
