+++ src/contents/_domain/interfaces.ts
