+++ src/contents/_domain/processors/ImageProcessor.ts
